NG-103 <매물 삭제 기능>
